+++ packages/renderer/src/main.ts
@@ -1,7 +1,0 @@
-import { createApp } from 'vue'
-import App from './App.vue'
-import './samples/node-api'
-
-createApp(App)
-  .mount('#app')
-  .$nextTick(window.removeLoading)