+++ src/main.ts
@@ -1,0 +1,7 @@
+import { createApp } from 'vue'
+import App from './App.vue'
+import './samples/node-api'
+
+createApp(App)
+  .mount('#app')
+  .$nextTick(window.removeLoading)